refactor: class data 타입 공통 타입 폴더로 분리
